bluetooth-fw/nimble: add support for bonding storage
Adds an implementation to store bonding data required to re-establish
encrypted connections. As you'll see, only a limited set of all the
NimBLE features are supported (the ones we care about), so if ever
reconfiguring NimBLE with other settings this implementation may NOT be
sufficient for things to work. As of today, we just store LTK (common in
peer/our storage) and IRK (to resolve peer addresses).

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>