asterix: add some fake OTP / serial data until we have a real mfg flow
The iOS app gets quite upset if there is no serial number or MFG data,
and the current fake OTP data incorrectly reports that it is programmed
already.  Later, we should implement OTP in terms of UICR (FIRM-35) but
for now at least we should not report it as programmed if it is not.

Signed-off-by: Joshua Wise <joshua@accelerated.tech>