bluetooth-fw/nimble: add support for store gen key callback
This allows NimBLE to always use the same, randomly generated, IRK if
RPA is ever used. LTK generation is not supported, this way, NimBLE will
generate a random LTK every time, not based on any root key.

Note that ble_root_keys_get_and_generate_if_needed is not called when
using NimBLE because the entropy source we use is the controller one
(via ble_hs_hci_rand), so it can't be accessed before the
host/controller are running.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>