platform/asterix: kick watchdog if enabled
On nRF52840, WDT cannot be disabled once started. It even survives a
system reset, so the bootloader must kick it on operations that can
take >= WDT expiration time, now configured at 8s. This mostly happens
during FW upgrades, where flash erase/write takes time. This patch is a
bit hacky, ideally we should just have a timer that kicks it periodically
as we do not pretend to protect the bootloader from deadlocks.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>