fw/nrf52: Use UICR to simulate OTP on nRF5 platforms. [FIRM-35]
nRF5 doesn't actually have true one-time programmable memory, but it
does have some memory that is separate from the user program and some
memory that is separate from the external SPI flash, called UICR.  UICR
can still be erased, and it is vaguely possible to accidentally erase it
with OpenOCD's nrf52_recover command... but all that said, this is
probably the most reasonable place to put serial numbers.  (Pebble
serial numbers are neither read-protected nor forgery-proof: they are
not signed at manufacture.  So if a user wants to mess with their
hardware in a way that might need nrf52_recover'ing, then it may not be
a bad idea in the future to provide a tool in this repository to back up
and restore the UICR.)

Tested by building PRF and using 'serial read', 'pcbaserial write',
'hwver write' commands.  Note that since asterix is still treated as
IS_BIGBOARD, main.c tries to assign it a serial number on boot!

Signed-off-by: Joshua Wise <joshua@accelerated.tech>