bluetooth-fw/nimble: allow repeat pairing while in PRF
If a phone deletes pairing information and watch is left in PRF mode
we won't be able to connect again because NimBLE doesn't handle the
repeat pairing event. While in normal firmware, we can always go to BT
settings menu and delete a pairing, we cannot in PRF. Ideally we should
warn user about this event. This is left as a TODO for now.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>