fw/services/battery: do not trigger LP/critical states if plugged
Regardless of the battery state, stay in good state if the watch is
plugged. This can happen e.g. in development where the battery is in
reality a power source that can be turned off if operating from VUSB
only. In a real watch, this will just happen if battery is completely
dead, but still, allow using the watch if plugged. Nordic PMIC will by
default not even try to charge a battery like that. If this is a safety
concern for older designs this can be made configurable.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>