fw/kernel/util/standby: disable display
Disable display after clearing. This is really added because on Asterix,
when PMIC is fully shutdown there seems to be a glitch that enables
display again, so showing what the display has on its memory (a partial
content, apparently erase does not fully wipe internal memory?).

Asterix should probably pull the DISP line, or we should analyze what is
the root cause of this problem. However, a display disable seems to
prevent this to happen, and change is pretty harmless.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>