src/fw/kernel: enable icache on nRF52, except when asleep
The NVMC on nRF52 has an instruction cache that is disabled by default.
We would like to have it turned on when the CPU is running -- this
improves performance, but more to the point, it improves performance per
watt while running (and means that we can shut down the CPU sooner).
Turn it on during init, and then turn it off when we go to sleep (it
costs us as much as 200 uA of sleep current, otherwise!).

Signed-off-by: Joshua Wise <joshua@accelerated.tech>