fw/kernel/util/stop: WFI: Disable entering deepsleep mode
Disable configuring the processor core to enter deepsleep mode when
executing WFI. If the PLL is not re-enabled quickly enough after
exiting deepsleep, the BLUETOOTH_UART may not be ready in time to
receive data from the cc2564, leading to data loss (e.g., missing
RTS transitions).

This change is a temporary workaround and should be reverted once
proper EHCILL Four-Wire Power Management protocol support for cc2564
is implemented.

Signed-off-by: Magdalena Kasenberg <magdalena.kasenberg@codecoup.pl>