waftools/protoc: fix race condition in protobuf generation
nanopb_generator will attempt to use protoc to generate a Python file as part of
it's operation, if that file doesn't exist yet. It seems like when waf runs
those generations in parallel, there's a race to generate that file that
results in build failures when the race is lost. To resolve this, we call the
function that generates that file during configure so we avoid the parallel
race during build.

Signed-off-by: Ruby Iris Juric <ruby@srxl.me>