fw/drivers/sf32lb52/jdi: add hack to mitigate crash due to race
We have some sort of race while updating the display (we enter with
critical section unset, ie, !vPortInCritical) but at the time we unlock
the mutex, it appears we are. Adding a sad delay seems to, at least,
make it significantly less frequent.

This needs to be reverted once we find the root cause of the problem.
Just adding this so system is less crashy and we can continue testing
other things.

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>