bluetooth-fw/nimble_store: pretend to persist CCCDs
On iOS, on the *second* connection attempt to a cc256x device from the
legacy app, the PPoG session will never wake up; we'll see an attempted
write to the GATT Service Changed CCCD, but we'll respond with an
"Insufficient Authorization (0x08)".  What on earth?  Nobody even references
BLE_ATT_ERR_INSUFFICIENT_AUTHOR!  But, it turns out, 0x08 is shared with
BLE_HS_ENOTSUP, and we *can* end up returning that from the nimble_store
code by way of `ble_gatts_clt_cfg_access` -- and this only happens on the
*second* time, because, according to `ble_gatts_clt_cfg_access_locked`,
"Successful writes get persisted for bonded connections.".  When this
happens, unrelated things go wrong later (like PPoG, for some reason).

Well, we cheat: we don't actually persist it, since our store is not complex
enough to persist it to.  But nobody cares, so it's OK.

Fixes #260.

Signed-off-by: Joshua Wise <joshua@joshuawise.com>